split out tags API to /forum/tags endpoint
